requirements.txt
numpy
opencv-python
pyzbar
paho-mqtt
PyJWT 
mosquitto-clients